+++ marshmallow/schema.py
@@ -776,7 +776,7 @@ class BaseSchema(base.SchemaABC):
             if many:
                 for idx, item in enumerate(data):
                     try:
-                        value = item[field_name]
+                        value = item[field_obj.attribute or field_name]
                     except KeyError:
                         pass
                     else:
@@ -789,7 +789,7 @@ class BaseSchema(base.SchemaABC):
                         )
             else:
                 try:
-                    value = data[field_name]
+                    value = data[field_obj.attribute or field_name]
                 except KeyError:
                     pass
                 else: